Create database connection

+++ src/main/java/com/example/payment/strategy/IPaymentStrategy.java
@@ -1,9 +1,11 @@
 package com.example.payment.strategy;
 
+import java.math.BigDecimal;
+
 public interface IPaymentStrategy{
     String getType();
 
-    default String process(int amount) {
+    default String process(BigDecimal amount) {
         return amount + " TL " + getType() + " ile ödendi.";
     }
 }
